+++ src/TaskRunner/Commands/DrupalCommands.php
@@ -8,6 +8,10 @@ use Composer\Semver\Semver;
 use Consolidation\AnnotatedCommand\CommandData;
 use EcEuropa\Toolkit\TaskRunner\AbstractCommands;
 use EcEuropa\Toolkit\Toolkit;
+use EcEuropa\Toolkit\Website;
+use Robo\Contract\VerbosityThresholdInterface;
+use Robo\ResultData;
+use Robo\Symfony\ConsoleIO;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Yaml\Yaml;
 
@@ -143,7 +147,7 @@ class DrupalCommands extends AbstractCommands
 
         $tasks = [
             // Note that the chmod() method takes decimal values.
-            $this->taskFilesystemStack()->chmod($subdirPath, octdec('775'), 0000, true),
+            $this->taskFilesystemStack()->chmod($subdirPath, octdec('775')),
         ];
 
         if (file_exists($subdirPath . '/settings.php')) {
@@ -508,6 +512,62 @@ class DrupalCommands extends AbstractCommands
         }
 
         return $qaCompatibilityResult;
+    }
+
+    /**
+     * Command to check the forbidden permissions.
+     *
+     * @param array $options
+     *   Command options.
+     *
+     * @command drupal:check-permissions
+     *
+     * @option endpoint The endpoint to use to connect to QA Website.
+     * @option blocker  If given and in case of error the command will fail.
+     *
+     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
+     * @SuppressWarnings(PHPMD.NPathComplexity)
+     */
+    public function drupalCheckPermissions(ConsoleIO $io, array $options = [
+        'endpoint' => InputOption::VALUE_REQUIRED,
+        'blocker' => InputOption::VALUE_NONE,
+    ]): int
+    {
+        $blocker = $options['blocker'] === true || $this->getConfig()->get('drupal.permissions.blocker');
+        if (!empty($options['endpoint'])) {
+            Website::setUrl($options['endpoint']);
+        }
+        if (empty($auth = Website::apiAuth())) {
+            return ResultData::EXITCODE_ERROR;
+        }
+        $data = Website::get(Website::url() . '/api/v1/forbidden-permissions', $auth);
+        $data = json_decode($data, true);
+        if (empty($data) || !isset($data['forbidden_permissions'])) {
+            return ResultData::EXITCODE_ERROR;
+        }
+        $forbiddenPermissions = $data['forbidden_permissions'];
+
+        $roles = $this->taskExec($this->getBin('drush'))
+            ->arg('role:list')
+            ->setVerbosityThreshold(VerbosityThresholdInterface::VERBOSITY_DEBUG)
+            ->run()->getMessage();
+        $roles = (array) Yaml::parse($roles);
+
+        $fail = false;
+        foreach ($roles as $role) {
+            foreach ($role['perms'] ?? [] as $permission) {
+                if (in_array($permission, $forbiddenPermissions)) {
+                    $io->say(sprintf(
+                        "The role '%s' contains a forbidden permission '%s'",
+                        $role['label'],
+                        $permission
+                    ));
+                    $fail = true;
+                }
+            }
+        }
+
+        return $fail && $blocker ? ResultData::EXITCODE_ERROR : ResultData::EXITCODE_OK;
     }
 
     /**